+++ src/forms/finalFormTextArea.jsx
@@ -3,35 +3,24 @@ import { FormGroup, HelpBlock, Col, FormControl } from 'patternfly-react';
 import PropTypes from 'prop-types';
 import { inputObjectProps, metaObjectProps } from './finalFormPropTypes';
 
-const FinalFormField = ({
-  input,
+export const FinalFormTextArea = ({
   meta,
-  placeholder,
+  input,
   label,
-  type,
 }) => (
   <FormGroup validationState={meta.error ? 'error' : null}>
     <Col xs={2} componentClass="label" className="control-label">
       {label}
     </Col>
     <Col xs={8}>
-      <FormControl type={type} {...input} placeholder={placeholder} />
+      <FormControl componentClass="textarea" {...input} />
       {meta.error && <HelpBlock>{meta.error}</HelpBlock>}
     </Col>
   </FormGroup>
 );
 
-FinalFormField.propTypes = {
+FinalFormTextArea.propTypes = {
   input: inputObjectProps,
   meta: metaObjectProps,
-  placeholder: PropTypes.string,
   label: PropTypes.string.isRequired,
-  type: PropTypes.string,
 };
-
-FinalFormField.defaultProps = {
-  placeholder: '',
-  type: 'text',
-};
-
-export default FinalFormField;
