+++ src/forms/index.js
@@ -1,0 +1,4 @@
+export * from './finalFormField';
+export * from './finalFormPropTypes';
+export * from './finalFormCheckBox';
+export * from './finalFormTextArea';
