+++ src/forms/tests/finalFormCheckBox.test.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import renderer from 'react-test-renderer';
-import FinalFormField from '../finalFormField';
+import { FinalFormCheckBox } from '../finalFormCheckBox';
 
-describe('Final form input component', () => {
+describe('Final form checkbox component', () => {
   const initialProps = {};
   beforeEach(() => {
     initialProps.meta = {
@@ -28,21 +28,21 @@ describe('Final form input component', () => {
   });
 
   it('Should render correctly', () => {
-    const tree = renderer.create(<FinalFormField {...initialProps} />).toJSON();
+    const tree = renderer.create(<FinalFormCheckBox {...initialProps} />).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('Should render with error message', () => {
     const props = { ...initialProps, meta: { ...initialProps.meta, error: 'Error message' } };
-    const tree = renderer.create(<FinalFormField {...props} />).toJSON();
+    const tree = renderer.create(<FinalFormCheckBox {...props} />).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('Should call onChange event', () => {
     const onChange = jest.fn();
     const props = { ...initialProps, input: { ...initialProps.input, onChange } };
-    const wrapper = mount(<FinalFormField {...props} />);
-    wrapper.find('input[type="text"]').simulate('change', { value: 'new value' });
+    const wrapper = mount(<FinalFormCheckBox {...props} />);
+    wrapper.find('input[type="checkbox"]').simulate('change');
     expect(onChange).toHaveBeenCalled();
   });
 });
