+++ src/forms/tests/finalFormField.test.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import renderer from 'react-test-renderer';
-import FinalFormTextArea from '../finalFormTextArea';
+import { FinalFormField } from '../finalFormField';
 
-describe('Final form textarea component', () => {
+describe('Final form input component', () => {
   const initialProps = {};
   beforeEach(() => {
     initialProps.meta = {
@@ -28,21 +28,21 @@ describe('Final form textarea component', () => {
   });
 
   it('Should render correctly', () => {
-    const tree = renderer.create(<FinalFormTextArea {...initialProps} />).toJSON();
+    const tree = renderer.create(<FinalFormField {...initialProps} />).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('Should render with error message', () => {
     const props = { ...initialProps, meta: { ...initialProps.meta, error: 'Error message' } };
-    const tree = renderer.create(<FinalFormTextArea {...props} />).toJSON();
+    const tree = renderer.create(<FinalFormField {...props} />).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('Should call onChange event', () => {
     const onChange = jest.fn();
     const props = { ...initialProps, input: { ...initialProps.input, onChange } };
-    const wrapper = mount(<FinalFormTextArea {...props} />);
-    wrapper.find('textarea').simulate('change', { value: 'new value' });
+    const wrapper = mount(<FinalFormField {...props} />);
+    wrapper.find('input[type="text"]').simulate('change', { value: 'new value' });
     expect(onChange).toHaveBeenCalled();
   });
 });
