+++ src/forms/tests/finalFormTextArea.test.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import renderer from 'react-test-renderer';
-import FinalFormCheckBox from '../finalFormCheckBox';
+import { FinalFormTextArea } from '../finalFormTextArea';
 
-describe('Final form checkbox component', () => {
+describe('Final form textarea component', () => {
   const initialProps = {};
   beforeEach(() => {
     initialProps.meta = {
@@ -28,21 +28,21 @@ describe('Final form checkbox component', () => {
   });
 
   it('Should render correctly', () => {
-    const tree = renderer.create(<FinalFormCheckBox {...initialProps} />).toJSON();
+    const tree = renderer.create(<FinalFormTextArea {...initialProps} />).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('Should render with error message', () => {
     const props = { ...initialProps, meta: { ...initialProps.meta, error: 'Error message' } };
-    const tree = renderer.create(<FinalFormCheckBox {...props} />).toJSON();
+    const tree = renderer.create(<FinalFormTextArea {...props} />).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('Should call onChange event', () => {
     const onChange = jest.fn();
     const props = { ...initialProps, input: { ...initialProps.input, onChange } };
-    const wrapper = mount(<FinalFormCheckBox {...props} />);
-    wrapper.find('input[type="checkbox"]').simulate('change');
+    const wrapper = mount(<FinalFormTextArea {...props} />);
+    wrapper.find('textarea').simulate('change', { value: 'new value' });
     expect(onChange).toHaveBeenCalled();
   });
 });
