+++ src/vm-snapshot-form/index.js
@@ -1,4 +1,1 @@
 export * from './snapshot-form/snapshotForm';
-export * from './finalFormField';
-export * from './finalFormTextArea';
-export * from './finalFormCheckBox';
